Fixing a SequentialDelay reconnect strategy bug, waitUntil must be set.
The constructor of SequentialDelayed was not setting the waitUntil value; it was only
paying attention to whether or not it was NULL. Whilst the primary case for adding this
constructor argument was for this purpose, there is no reason the value cannot be set
too. A unit test has been added which highlights the issue.